+++ Optuna/Pruner/IPruner.cs
@@ -1,0 +1,7 @@
+namespace Optuna.Pruner
+{
+    public interface IPruner
+    {
+        dynamic ToPython(dynamic optuna);
+    }
+}