+++ Tunny/Solver/Algorithm.cs
@@ -1,12 +1,16 @@
 using System;
 using System.Collections.Generic;
 using System.Globalization;
+using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading;
 using System.Windows;
 
+using Optuna.Dashboard.HumanInTheLoop;
 using Optuna.Study;
+using Optuna.Util;
 
 using Python.Runtime;
 
@@ -103,7 +107,7 @@ namespace Tunny.Solver
         private void PreferentialOptimization(int nBatch, dynamic storage, dynamic artifactBackend, out Parameter[] parameter, out TrialGrasshopperItems result, out dynamic study)
         {
             TLog.MethodStart();
-            var preferentialOpt = new HumanInTheLoop.Preferential(TEnvVariables.TmpDirPath, Settings.Storage.Path);
+            var preferentialOpt = new Preferential(TEnvVariables.TmpDirPath, Settings.Storage.Path);
             if (Objective.Length > 1)
             {
                 TunnyMessageBox.Show("Human-in-the-Loop(Preferential GP optimization) only supports single objective optimization. Optimization is run without considering constraints.", "Tunny");
@@ -112,7 +116,7 @@ namespace Tunny.Solver
             study = preferentialOpt.CreateStudy(nBatch, Settings.StudyName, storage, objNickName[0]);
             var optInfo = new OptimizationHandlingInfo(int.MaxValue, 0, study, storage, artifactBackend, FishEgg, objNickName);
             SetStudyUserAttr(study, PyConverter.EnumeratorToPyList(Variables.Select(v => v.NickName)), false);
-            HumanInTheLoop.HumanInTheLoopBase.WakeOptunaDashboard(Settings.Storage);
+            HumanInTheLoopBase.WakeOptunaDashboard(Settings.Storage.Path, TEnvVariables.PythonPath);
             optInfo.Preferential = preferentialOpt;
             RunPreferentialOptimize(optInfo, nBatch, out parameter, out result);
         }
@@ -121,7 +125,8 @@ namespace Tunny.Solver
         {
             TLog.MethodStart();
             PyObject optuna = Py.Import("optuna");
-            study = Study.CreateStudy(optuna, Settings.StudyName, sampler, directions, storage, Settings.Optimize.ContinueStudy);
+            dynamic pruner = Settings.Pruner.ToPython();
+            study = Study.CreateStudy(optuna, Settings.StudyName, sampler, directions, storage, pruner, Settings.Optimize.ContinueStudy);
             string[] objNickName = Objective.GetNickNames();
             var optInfo = new OptimizationHandlingInfo(nTrials, timeout, study, storage, artifactBackend, FishEgg, objNickName);
             SetStudyUserAttr(study, PyConverter.EnumeratorToPyList(Variables.Select(v => v.NickName)));
@@ -132,12 +137,12 @@ namespace Tunny.Solver
         {
             TLog.MethodStart();
             PyObject optuna = Py.Import("optuna");
-            study = Study.CreateStudy(optuna, Settings.StudyName, sampler, directions, storage, Settings.Optimize.ContinueStudy);
+            study = Study.CreateStudy(optuna, Settings.StudyName, sampler, directions, storage, null, Settings.Optimize.ContinueStudy);
             string[] objNickName = Objective.GetNickNames();
             var optInfo = new OptimizationHandlingInfo(int.MaxValue, timeout, study, storage, artifactBackend, FishEgg, objNickName);
             SetStudyUserAttr(study, PyConverter.EnumeratorToPyList(Variables.Select(v => v.NickName)));
-            var humanSliderInput = new HumanInTheLoop.HumanSliderInput(TEnvVariables.TmpDirPath, Settings.Storage.Path);
-            HumanInTheLoop.HumanInTheLoopBase.WakeOptunaDashboard(Settings.Storage);
+            var humanSliderInput = new HumanSliderInput(TEnvVariables.TmpDirPath, Settings.Storage.Path);
+            HumanInTheLoopBase.WakeOptunaDashboard(Settings.Storage.Path, TEnvVariables.PythonPath);
             humanSliderInput.SetObjective(study, objNickName);
             humanSliderInput.SetWidgets(study, objNickName);
             optInfo.HumanSliderInput = humanSliderInput;
@@ -253,7 +258,7 @@ namespace Tunny.Solver
                 {
                     break;
                 }
-                if (HumanInTheLoop.Preferential.GetRunningTrialNumber(optInfo.Study) >= nBatch)
+                if (Preferential.GetRunningTrialNumber(optInfo.Study) >= nBatch)
                 {
                     continue;
                 }
@@ -279,7 +284,7 @@ namespace Tunny.Solver
                 {
                     break;
                 }
-                if (HumanInTheLoop.HumanSliderInput.GetRunningTrialNumber(optInfo.Study) >= nBatch)
+                if (HumanSliderInput.GetRunningTrialNumber(optInfo.Study) >= nBatch)
                 {
                     continue;
                 }
@@ -293,6 +298,7 @@ namespace Tunny.Solver
         private TrialGrasshopperItems RunSingleOptimizeStep(OptimizationHandlingInfo optInfo, Parameter[] parameter, int trialNum, DateTime startTime)
         {
             TLog.MethodStart();
+            dynamic optuna = Py.Import("optuna");
             dynamic trial;
             int progress = trialNum * 100 / optInfo.NTrials;
             var result = new TrialGrasshopperItems();
@@ -309,7 +315,7 @@ namespace Tunny.Solver
 
                 trial = optInfo.Study.ask();
                 SetOptimizationParameter(parameter, trial);
-                ProgressState pState = SetProgressState(optInfo, parameter, trialNum, startTime);
+                ProgressState pState = SetProgressState(optInfo, parameter, trialNum, startTime, trial, Settings.Pruner);
                 if (Settings.Optimize.IgnoreDuplicateSampling && IsSampleDuplicate(trial, out result))
                 {
                     TLog.Info($"Trial {trialNum} is duplicate sample.");
@@ -353,9 +359,13 @@ namespace Tunny.Solver
                     result.Artifacts.UploadArtifacts(optInfo.ArtifactBackend, trial);
                 }
 
-                if (result.Attribute.TryGetValue("IsFAIL", out List<string> isFail) && isFail.Contains("True"))
+                if (!(bool)optInfo.Study._is_multi_objective() && (bool)trial.should_prune())
                 {
-                    dynamic optuna = Py.Import("optuna");
+                    optInfo.Study.tell(trial, state: optuna.trial.TrialState.PRUNED);
+                    TLog.Warning($"Trial {trialNum} pruned.");
+                }
+                else if (result.Attribute.TryGetValue("IsFAIL", out List<string> isFail) && isFail.Contains("True"))
+                {
                     optInfo.Study.tell(trial, state: optuna.trial.TrialState.FAIL);
                     TLog.Warning($"Trial {trialNum} failed.");
                 }
@@ -382,20 +392,8 @@ namespace Tunny.Solver
         {
             double[] values;
             PyModule ps = Py.CreateScope();
-            ps.Exec(
-                "def check_duplicate(trial):\n" +
-                "    import optuna\n" +
-                "    from optuna.trial import TrialState\n" +
-                "    states_to_consider = (TrialState.COMPLETE,)\n" +
-                "    trials_to_consider = trial.study.get_trials(deepcopy=False, states=states_to_consider)\n" +
-                "    for t in reversed(trials_to_consider):\n" +
-                "        if trial.params == t.params:\n" +
-                "            trial.set_user_attr('NOTE', f'trial {t.number} and trial {trial.number} were duplicate parameters.')\n" +
-                "            if 'Constraint' in t.user_attrs:\n" +
-                "               trial.set_user_attr('Constraint', t.user_attrs['Constraint'])\n" +
-                "            return t.values\n" +
-                "    return None\n"
-            );
+            var assembly = Assembly.GetExecutingAssembly();
+            ps.Exec(ReadFileFromResource.Text(assembly, "Tunny.Solver.Python.check_duplication.py"));
             dynamic checkDuplicate = ps.Get("check_duplicate");
             values = checkDuplicate(trial);
             result = new TrialGrasshopperItems(values);
@@ -465,7 +463,18 @@ namespace Tunny.Solver
             int nTrials = optInfo.NTrials;
             double timeout = optInfo.Timeout;
             dynamic study = optInfo.Study;
-            bool studyStopFlag = study._stop_flag;
+            bool studyStopFlag = false;
+            if (optInfo.HumanSliderInput == null && optInfo.Preferential == null)
+            {
+                studyStopFlag = study._stop_flag;
+            }
+
+
+            if (File.Exists(TEnvVariables.QuitFishingPath))
+            {
+                OptimizeLoop.IsForcedStopOptimize = true;
+                File.Delete(TEnvVariables.QuitFishingPath);
+            }
 
             bool isOptimizeCompleted = false;
             if (trialNum >= nTrials)
@@ -493,7 +502,7 @@ namespace Tunny.Solver
             return isOptimizeCompleted;
         }
 
-        private ProgressState SetProgressState(OptimizationHandlingInfo optSet, Parameter[] parameter, int trialNum, DateTime startTime)
+        private ProgressState SetProgressState(OptimizationHandlingInfo optSet, Parameter[] parameter, int trialNum, DateTime startTime, dynamic trial, Pruner pruner)
         {
             TLog.MethodStart();
             double[][] bestValues = ComputeBestValues(optSet.Study);
@@ -504,6 +513,8 @@ namespace Tunny.Solver
                 BestValues = bestValues,
                 Parameter = parameter,
                 HypervolumeRatio = 0,
+                OptunaTrial = trial,
+                Pruner = pruner,
                 EstimatedTimeRemaining = optSet.Timeout <= 0
                     ? TimeSpan.FromSeconds((DateTime.Now - startTime).TotalSeconds * (optSet.NTrials - trialNum) / (trialNum + 1))
                     : TimeSpan.FromSeconds(optSet.Timeout - (DateTime.Now - startTime).TotalSeconds)